src/lab_05/docchat/pyproject.toml
[project]
name = "docchat"
version = "0.1.0"
description = "Document Chat Application with Azure AI integration"
authors = [{name = "DocChat Team"}]
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asgiref==3.8.1",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "beautifulsoup4==4.12.3",
    "build==1.2.2.post1",
    "cachetools==5.5.1",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.1.8",
    "cohere==5.13.11",
    "coloredlogs==15.0.1",
    "cryptography==44.0.0",
    "dataclasses-json==0.6.7",
    "deepdiff==8.1.1",
    "deepsearch-glm==1.0.0",
    "Deprecated==1.2.18",
    "dill==0.3.9",
    "distro==1.9.0",
    "docling==2.15.0",
    "docling-core==2.16.0",
    "docling-parse==3.1.2",
    "durationpy==0.9",
    "easyocr==1.7.2",
    "emoji==2.14.1",
    "et_xmlfile==2.0.0",
    "fastapi==0.115.7",
    "fastavro==1.10.0",
    "ffmpy==0.5.0",
    "filelock==3.17.0",
    "filetype==1.2.0",
    "flatbuffers==25.1.24",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.66.0",
    "gradio==5.13.2",
    "gradio_client==1.6.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "h11==0.14.0",
    "html5lib==1.1",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.28.0",
    "humanfriendly==10.0",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib_metadata==8.5.0",
    "importlib_resources==6.5.2",
    "Jinja2==3.1.5",
    "jiter==0.8.2",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonlines==3.1.0",
    "jsonpatch==1.33",
    "jsonpath-python==1.0.6",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kubernetes==32.0.0",
    "langchain==0.3.16",
    "langchain-cohere==0.4.2",
    "langchain-community==0.3.16",
    "langchain-core==0.3.32",
    "langchain-openai==0.3.2",
    "langchain-text-splitters==0.3.5",
    "langdetect==1.0.9",
    "langgraph==0.2.68",
    "langgraph-checkpoint==2.0.10",
    "langgraph-sdk==0.1.51",
    "langsmith==0.3.2",
    "lazy_loader==0.4",
    "loguru==0.7.3",
    "lomond==0.3.3",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "marko==2.1.2",
    "MarkupSafe==2.1.5",
    "marshmallow==3.26.0",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpire==2.10.2",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "multiprocess==0.70.17",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "ninja==1.11.1.3",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "oauthlib==3.2.2",
    "olefile==0.47",
    "onnxruntime==1.20.1",
    "openai==1.60.2",
    "opencv-python-headless==4.11.0.86",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.29.0",
    "opentelemetry-exporter-otlp-proto-common==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation==0.50b0",
    "opentelemetry-instrumentation-asgi==0.50b0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-proto==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-semantic-conventions==0.50b0",
    "opentelemetry-util-http==0.50b0",
    "orderly-set==5.2.3",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.1.4",
    "pillow==10.4.0",
    "posthog==3.11.0",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pyclipper==1.3.0.post6",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "pydantic_core==2.27.2",
    "pydub==0.25.1",
    "Pygments==2.19.1",
    "pypdf==5.2.0",
    "pypdfium2==4.30.1",
    "PyPika==0.48.9",
    "pyproject_hooks==1.2.0",
    "python-bidi==0.6.3",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-iso639==2025.1.28",
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "python-oxmsg==0.0.1",
    "python-pptx==1.0.2",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "rank-bm25==0.2.2",
    "RapidFuzz==3.11.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "Rtree==1.3.0",
    "ruff==0.9.3",
    "safehttpx==0.1.6",
    "safetensors==0.5.2",
    "scikit-image==0.25.1",
    "scipy==1.15.1",
    "semantic-version==2.10.0",
    "semchunk==2.2.2",
    "shapely==2.0.6",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "SQLAlchemy==2.0.37",
    "starlette==0.45.3",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "tenacity==9.0.0",
    "tifffile==2025.1.10",
    "tiktoken==0.8.0",
    "tokenizers==0.21.0",
    "tomlkit==0.13.2",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "transformers==4.48.1",
    "triton==3.2.0",
    "typer==0.12.5",
    "types-PyYAML==6.0.12.20241230",
    "types-requests==2.32.0.20241016",
    "typing-inspect==0.9.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "unstructured==0.16.17",
    "unstructured-client==0.25.9",
    "urllib3==2.1.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "XlsxWriter==3.2.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
    "pysqlite3-binary==0.5.4",
    "azure-ai-inference==1.0.0b6",
    "azure-identity==1.21.0",
    "azure-core",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
docchat = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "document_processor", "retriever", "utils"]
include = ["app.py"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true